Rename shared model AdaptiveCard files
