+++ source/android/adaptivecards/src/main/java/com/microsoft/adaptivecards/objectmodel/AdaptiveCardSchemaKey.java
@@ -39,6 +39,7 @@ public final class AdaptiveCardSchemaKey {
   public final static AdaptiveCardSchemaKey Container = new AdaptiveCardSchemaKey("Container");
   public final static AdaptiveCardSchemaKey ContainerStyleConfig = new AdaptiveCardSchemaKey("ContainerStyleConfig");
   public final static AdaptiveCardSchemaKey Dark = new AdaptiveCardSchemaKey("Dark");
+  public final static AdaptiveCardSchemaKey Data = new AdaptiveCardSchemaKey("Data");
   public final static AdaptiveCardSchemaKey DateInput = new AdaptiveCardSchemaKey("DateInput");
   public final static AdaptiveCardSchemaKey Default = new AdaptiveCardSchemaKey("Default");
   public final static AdaptiveCardSchemaKey Emphasis = new AdaptiveCardSchemaKey("Emphasis");
@@ -147,7 +148,7 @@ public final class AdaptiveCardSchemaKey {
     swigNext = this.swigValue+1;
   }
 
-  private static AdaptiveCardSchemaKey[] swigValues = { Accent, ActionAlignment, ActionMode, ActionOrientation, Actions, ActionSetConfig, ActionsOrientation, AdaptiveCard, AltText, Attention, BackgroundColor, BackgroundImageUrl, BaseCardElement, Body, BorderColor, BorderThickness, ButtonSpacing, Card, Center, Choices, ChoiceSet, Color, ColorConfig, Colors, Column, Columns, ColumnSet, Container, ContainerStyleConfig, Dark, DateInput, Default, Emphasis, ExtraLarge, Facts, FactSet, FallbackText, FontFamily, FontSizes, Good, HorizontalAlignment, Id, Image, Images, ImageSet, ImageSize, ImageSizes, InlineTopMargin, IsMultiline, IsMultiSelect, IsRequired, IsSelected, IsSubtle, Items, Large, Left, Light, LineColor, LineThickness, Max, MaxActions, MaxLength, MaxLines, Medium, Method, Min, MinVersion, Normal, NumberInput, Padding, Placeholder, Right, Separation, ShowActionMode, ShowCard, ShowCardActionConfig, Size, Small, Spacing, SpacingDefinition, Speak, Stretch, StrongSeparation, Style, Subtle, SupportsInteractivity, Text, TextBlock, TextConfig, TextInput, TextWeight, TimeInput, Title, ToggleInput, Type, Url, Value, ValueOff, ValueOn, Version, Warning, Weight, Wrap };
+  private static AdaptiveCardSchemaKey[] swigValues = { Accent, ActionAlignment, ActionMode, ActionOrientation, Actions, ActionSetConfig, ActionsOrientation, AdaptiveCard, AltText, Attention, BackgroundColor, BackgroundImageUrl, BaseCardElement, Body, BorderColor, BorderThickness, ButtonSpacing, Card, Center, Choices, ChoiceSet, Color, ColorConfig, Colors, Column, Columns, ColumnSet, Container, ContainerStyleConfig, Dark, Data, DateInput, Default, Emphasis, ExtraLarge, Facts, FactSet, FallbackText, FontFamily, FontSizes, Good, HorizontalAlignment, Id, Image, Images, ImageSet, ImageSize, ImageSizes, InlineTopMargin, IsMultiline, IsMultiSelect, IsRequired, IsSelected, IsSubtle, Items, Large, Left, Light, LineColor, LineThickness, Max, MaxActions, MaxLength, MaxLines, Medium, Method, Min, MinVersion, Normal, NumberInput, Padding, Placeholder, Right, Separation, ShowActionMode, ShowCard, ShowCardActionConfig, Size, Small, Spacing, SpacingDefinition, Speak, Stretch, StrongSeparation, Style, Subtle, SupportsInteractivity, Text, TextBlock, TextConfig, TextInput, TextWeight, TimeInput, Title, ToggleInput, Type, Url, Value, ValueOff, ValueOn, Version, Warning, Weight, Wrap };
   private static int swigNext = 0;
   private final int swigValue;
   private final String swigName;
