+++ source/shared/cpp/ObjectModel/SharedAdaptiveCard.cpp
@@ -1,4 +1,4 @@
-#include "AdaptiveCard.h"
+#include "SharedAdaptiveCard.h"
 #include "ChoiceSetInput.h"
 #include "ColumnSet.h"
 #include "Container.h"
