+++ source/shared/cpp/ObjectModel/SharedAdaptiveCard.h
