+++ source/shared/cpp/ObjectModel/ShowCardAction.cpp
@@ -1,4 +1,4 @@
-#include "AdaptiveCard.h"
+#include "SharedAdaptiveCard.h"
 #include "ParseUtil.h"
 #include "ShowCardAction.h"
