+++ source/uwp/ObjectModelProjection/ObjectModelProjection.vcxproj
@@ -46,7 +46,7 @@
     <ClCompile Include="..\..\shared\cpp\ObjectModel\ToggleInput.cpp" />
     <ClCompile Include="..\..\shared\cpp\ObjectModel\OpenUrlAction.cpp" />
     <ClCompile Include="..\..\shared\cpp\ObjectModel\ParseUtil.cpp" />
-    <ClCompile Include="..\..\shared\cpp\ObjectModel\AdaptiveCard.cpp" />
+    <ClCompile Include="..\..\shared\cpp\ObjectModel\SharedAdaptiveCard.cpp" />
     <ClCompile Include="..\..\shared\cpp\ObjectModel\BaseCardElement.cpp" />
     <ClCompile Include="..\..\shared\cpp\ObjectModel\Container.cpp" />
     <ClCompile Include="..\..\shared\cpp\ObjectModel\Enums.cpp" />
@@ -76,7 +76,7 @@
     <ClInclude Include="..\..\shared\cpp\ObjectModel\ToggleInput.h" />
     <ClInclude Include="..\..\shared\cpp\ObjectModel\OpenUrlAction.h" />
     <ClInclude Include="..\..\shared\cpp\ObjectModel\ParseUtil.h" />
-    <ClInclude Include="..\..\shared\cpp\ObjectModel\AdaptiveCard.h" />
+    <ClInclude Include="..\..\shared\cpp\ObjectModel\SharedAdaptiveCard.h" />
     <ClInclude Include="..\..\shared\cpp\ObjectModel\BaseCardElement.h" />
     <ClInclude Include="..\..\shared\cpp\ObjectModel\Container.h" />
     <ClInclude Include="..\..\shared\cpp\ObjectModel\Enums.h" />
