+++ source/uwp/ObjectModelProjection/ObjectModelProjection.vcxproj.filters
@@ -1,7 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClCompile Include="..\..\shared\cpp\ObjectModel\AdaptiveCard.cpp" />
     <ClCompile Include="..\..\shared\cpp\ObjectModel\Container.cpp" />
     <ClCompile Include="..\..\shared\cpp\ObjectModel\Enums.cpp" />
     <ClCompile Include="..\..\shared\cpp\ObjectModel\TextBlock.cpp" />
@@ -30,9 +29,9 @@
     <ClCompile Include="..\..\shared\cpp\ObjectModel\TextInput.cpp" />
     <ClCompile Include="..\..\shared\cpp\ObjectModel\TimeInput.cpp" />
     <ClCompile Include="..\..\shared\cpp\ObjectModel\ToggleInput.cpp" />
+    <ClCompile Include="..\..\shared\cpp\ObjectModel\SharedAdaptiveCard.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\shared\cpp\ObjectModel\AdaptiveCard.h" />
     <ClInclude Include="..\..\shared\cpp\ObjectModel\Container.h" />
     <ClInclude Include="..\..\shared\cpp\ObjectModel\Enums.h" />
     <ClInclude Include="..\..\shared\cpp\ObjectModel\TextBlock.h" />
@@ -66,6 +65,7 @@
     <ClInclude Include="..\..\shared\cpp\ObjectModel\TimeInput.h" />
     <ClInclude Include="..\..\shared\cpp\ObjectModel\ToggleInput.h" />
     <ClInclude Include="..\..\shared\cpp\ObjectModel\HostConfig.h" />
+    <ClInclude Include="..\..\shared\cpp\ObjectModel\SharedAdaptiveCard.h" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="json">
